Handle missing notify capability in 8 port devices
The command gpio notify get is not available in 8 port devices.
Thus the response to a 'gpio notify get' command is empty as
opposed to all other devices.

Unit tests have been extended to cover the different devices and
also the missing notification feature for 8 port devices is modeled
in the device mockup.