Re-organize queries
With the goal to further simplify the code, this commit applies the
single responsibility principle to the query method. This is done by
removing the validation extra-functionality into a read_response method
which is now called after the query. The read_response method accepts
the expected response as optional argument so that the response can be
validated in a uniform way.

Additionally resolved several issues reported by static code analysis.