Ignore error when closing an already closed serial connection
An OSError would be raised when reader and writer threads were
concurrently closing the serial connection. One of them would close
an already closed connection.

Ideally this would be synchronized so we'd be able to detect an issue
closing the connection. In this case a global resource in the system may
be broken rendering the system un-usable.

For now we just ignore any issues while closing, not taking a potential
later use of the device into account.