Mitigate flakyness in notification tests
A race condition in the test, where a notification is injected at the
end, i.e. after the response. So the response is returned to the main
thread while callbacks are still being called in the polling thread.

In real-life applications no assumption can be made as to when a
notification is received, so this is a test-only issue.